Update the graph api and add a client

+++ modules/model/src/main/scala/lucuma/itc/math.scala
@@ -3,17 +3,17 @@
 
 package lucuma.itc.math
 
-import scala.math._
+import scala.math.*
 
 // Round the value to n significant figures, this to remove
 // the extra precision produced by ITC calculations using doubles
 // This will not work for extremely small numbers but we don't expect such
-def roundToSignificantFigures(num: Double, n: Int): Double =
+def roundToSignificantFigures(num: BigDecimal, n: Int): Double =
   if num == 0 then 0
   else
-    val d     = ceil(log10(abs(num)))
+    val d     = ceil(log10(abs(num.toDouble)))
     val power = n - d.toInt
 
     val magnitude = pow(10, power)
-    val shifted   = round(num * magnitude)
+    val shifted   = round(num.toDouble * magnitude)
     shifted / magnitude
